Se ajusta excepción global, se genera el readme para saber temas del proyecto, apis, como subirlo y demas, y se arregla puerto para BD a 3308 ya que este es el que manejo yo.

+++ backend/src/main/java/co/com/walmart/stefanini/prueba/PruebaWalmart/service/impl/PersonServiceImpl.java
@@ -1,13 +1,14 @@
 package co.com.walmart.stefanini.prueba.PruebaWalmart.service.impl;
 
-import java.util.List;
-
-import org.springframework.stereotype.Service;
-
 import co.com.walmart.stefanini.prueba.PruebaWalmart.entity.Person;
 import co.com.walmart.stefanini.prueba.PruebaWalmart.repository.PersonRepository;
 import co.com.walmart.stefanini.prueba.PruebaWalmart.service.PersonService;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
+import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
+
+import java.util.List;
 
 @Service
 @RequiredArgsConstructor
@@ -27,22 +28,35 @@ public class PersonServiceImpl implements PersonService {
 
     @Override
     public Person update(Long id, Person person) {
-        Person existing = repository.findById(id).orElseThrow();
+        Person existing = repository.findById(id)
+            .orElseThrow(() -> new ResponseStatusException(
+                HttpStatus.NOT_FOUND, "Persona con ID " + id + " no encontrada"
+            ));
+
         existing.setNombre(person.getNombre());
         existing.setApellido(person.getApellido());
         existing.setFechaNacimiento(person.getFechaNacimiento());
         existing.setPuesto(person.getPuesto());
         existing.setSueldo(person.getSueldo());
+
         return repository.save(existing);
     }
 
     @Override
     public void delete(Long id) {
+        if (!repository.existsById(id)) {
+            throw new ResponseStatusException(
+                HttpStatus.NOT_FOUND, "No se puede eliminar. Persona con ID " + id + " no existe"
+            );
+        }
         repository.deleteById(id);
     }
 
     @Override
     public Person getById(Long id) {
-        return repository.findById(id).orElseThrow();
+        return repository.findById(id)
+            .orElseThrow(() -> new ResponseStatusException(
+                HttpStatus.NOT_FOUND, "Persona con ID " + id + " no encontrada"
+            ));
     }
 }
